use off_t rather than size_t for read/write_count, doesn't overflow at 4GB, yes moving more than 4GB of data over asp is a bad idea

+++ include/atalk/asp.h
@@ -60,7 +60,7 @@ typedef struct ASP {
     char                cmdbuf[ASP_CMDMAXSIZ];
     char                data[ASP_DATAMAXSIZ];  
     size_t		cmdlen, datalen;
-    size_t 		read_count, write_count;
+    off_t 		read_count, write_count;
 } *ASP;
 
 #define ASPFL_SLS	1
